remove schedule page demo, port query over to AdminSchedulePostingPage
